Fix issue when unfav a podcast without any episode in disk

+++ app/src/main/java/com/tapura/podmorecasts/download/DownloadUtils.java
@@ -41,8 +41,10 @@ public class DownloadUtils {
         if (manager != null) {
             if (epiIndex == -1) { // cancelAll
                 long[] ids = mRepository.getAllId(feed);
-                manager.remove(ids);
-                mRepository.remove(ids);
+                if (ids != null && ids.length > 0) {
+                    manager.remove(ids);
+                    mRepository.remove(ids);
+                }
             } else {
                 long id = mRepository.getId(feed, epiIndex);
                 manager.remove(id);
